backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
selenium==4.15.2
beautifulsoup4==4.12.2
google-generativeai==0.3.2
python-pptx==0.6.22
requests==2.31.0
python-dotenv==1.0.0
Pillow>=3.3.2,<=9.5.0
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2
lxml==4.9.3
webdriver-manager==4.0.1